Rename i_gui_context_tests to i_event_loop_tests
